Move spectrogram transforms
