+++ audiomentations/spec_augmentations/spec_frequency_mask.py
@@ -55,9 +55,9 @@ class SpecFrequencyMask(BaseSpectrogramTransform):
         if self.fill_mode == "mean":
             fill_value = np.mean(
                 magnitude_spectrogram[
-                    self.parameters["start_frequency_index"] : self.parameters[
-                        "end_frequency_index"
-                    ]
+                self.parameters["start_frequency_index"] : self.parameters[
+                    "end_frequency_index"
+                ]
                 ]
             )
         else:
@@ -65,26 +65,9 @@ class SpecFrequencyMask(BaseSpectrogramTransform):
             fill_value = self.fill_constant
         magnitude_spectrogram = magnitude_spectrogram.copy()
         magnitude_spectrogram[
-            self.parameters["start_frequency_index"] : self.parameters[
-                "end_frequency_index"
-            ]
+        self.parameters["start_frequency_index"] : self.parameters[
+            "end_frequency_index"
+        ]
         ] = fill_value
         return magnitude_spectrogram
 
-
-class SpecChannelShuffle(BaseSpectrogramTransform):
-    """
-    Shuffle the channels of a multichannel spectrogram (channels last).
-    This can help combat positional bias.
-    """
-    supports_multichannel = True
-    supports_mono = False
-
-    def randomize_parameters(self, magnitude_spectrogram):
-        super().randomize_parameters(magnitude_spectrogram)
-        if self.parameters["should_apply"]:
-            self.parameters["shuffled_channel_indexes"] = list(range(magnitude_spectrogram.shape[-1]))
-            random.shuffle(self.parameters["shuffled_channel_indexes"])
-
-    def apply(self, magnitude_spectrogram):
-        return magnitude_spectrogram[..., self.parameters["shuffled_channel_indexes"]]
